Conflicts:
	NeuralNetworkProject/src/neuralnetworkproject/Network.java
	NeuralNetworkProject/src/neuralnetworkproject/NeuralNetworkProject.java

Updated a lot of stuff, tried to implement back propagation, its still not fully functional.

+++ NeuralNetworkProject/src/neuralnetworkproject/TestPair.java
@@ -31,10 +31,11 @@ public class TestPair {
     public void setImageData(int[] imgDat){
         this.imageData = imgDat;
     }
-    public ArrayList<Double> getDataArrayList(){
-        ArrayList<Double> dat = new ArrayList<Double>();
+
+    public ArrayList<Integer> getDataArrayList(){
+        ArrayList<Integer> dat = new ArrayList<Integer>();
         for (int i = 0; i < this.imageData.length; i++) {
-            dat.add(new Double(this.imageData[i]));
+            dat.add(new Integer(this.imageData[i]));
         }
         return dat;
     }
